# code/webapp/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Waaiburg API",
        "description": "Waaiburg API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sections": {
            "get": {
                "tags": [
                    "sections"
                ],
                "summary": "Get list of sections",
                "description": "Returns list of sections",
                "operationId": "sectionsIndex",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {token}",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "sections": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Section"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "sections"
                ],
                "summary": "Create a new section",
                "description": "Returns the created section",
                "operationId": "sectionsStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Section"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Section created succesfully"
                                        },
                                        "section": {
                                            "$ref": "#/components/schemas/Section"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/sections/{id}": {
            "delete": {
                "tags": [
                    "sections"
                ],
                "summary": "Delete a section",
                "description": "Returns the deleted section",
                "operationId": "sectionsDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of section to return",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Section deleted succesfully"
                                        },
                                        "section": {
                                            "$ref": "#/components/schemas/Section"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "sections"
                ],
                "summary": "Update a section",
                "description": "Returns the updated section",
                "operationId": "sectionsUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of section to return",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Section"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Section updated succesfully"
                                        },
                                        "section": {
                                            "$ref": "#/components/schemas/Section"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Section": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Section 1"
                    },
                    "description": {
                        "type": "string",
                        "example": "Section 1"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-05-01 12:00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-05-01 12:00:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}